pyproject.toml
[tool.poetry]
name = "comps"
version = "0.1.0"
description = ""
authors = ["dr-neptune <mrosedatasci@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cytoolz = "^0.12.3"
pytest-subtests = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
